refactor(cli): extract classifiers command family into cli/commands/classifiers.ts
